Loads and correctly tints the icon for solved quizzes
Uses white for the done icon and the primary color
for the small category icon. Clean, nice and well documented.

Also includes an update to the build.gradle to be able to look at
sources form Studio.

Further fixes a NPE in AlphaPickerQuizView due to uninitialized
alphabet as well as lint warnings due to missing View constructors
within quiz views.

Change-Id: I70b2feaa015cd3d9b970499929a5d98cffbca111