Attempts to solve orientation changes by removing ViewPagers
The ViewPager used for display of quizzes is replaced by a
AdapterViewFlipper. There is no need to have pagination in a
ViewPager way. We only advance forward by tapping a FAB instead
of swiping between views.

Currently the height of the view is not optimal, but at least
we got rid of the ViewPager and it's dependencies.

Also the orientation changes are not completely propagated,
just like they were before. This has to be fixed in the future.

But there's also some more things in this CL:

No more crash when changing orientation due to
unreferenced fragment within quiz.
Also the toolbar elevation now matches the view's
elevation after an orientation change.
Each AbsQuizView now gets an id.
Remove unused methods.

Change-Id: I5f2efe75e156078445a1b052e915d8cd57dd9832