Animate solving of quizzes
When a quiz is solved a series of animations is
triggered. Starting from the fab and taking on the whole
view. Also some fine tuning for transitions during sign in
and the toolbar for the categories screen.
And we're no longer using a ViewPager, again.
Further there's less exported listener implementations.

Change-Id: I0d7dbea674a9123fef6dc120f589ba4fffc37fa1